WIP: Add RSQL
