--- src/main/java/eatseasyspring/eatseasyspring/model/RestFilterableData.java
@@ -1,60 +1,0 @@
-package eatseasyspring.eatseasyspring.model;
-
-
-
-import javax.persistence.Embeddable;
-
-import java.util.List;
-
-
-
-@Embeddable
-public class RestFilterableData {
-
-    private float gpsLat;
-    private float gpsLon;
-    private List<String> restTypes;
-
-    public List<String> getRestTypes() {
-        return restTypes;
-    }
-
-    public void setRestTypes(List<String> restTypes) {
-        this.restTypes = restTypes;
-    }
-
-
-
-/*    public RestFilterableData(float gpsLat, float gpsLon, List<String> restTypes) {
-        this.gpsLat = gpsLat;
-        this.gpsLon = gpsLon;
-       // this.restTypes = restTypes;
-    }
-
-    public RestFilterableData(RestFilterableData restFilterableData)
-    {
-        this.gpsLat = restFilterableData.gpsLat;
-        this.gpsLon = restFilterableData.gpsLon;
-     //   this.restTypes = restFilterableData.restTypes;
-    }*/
-
-    public void setGpsLat(float gpsLat) {
-        this.gpsLat = gpsLat;
-    }
-
-    public void setGpsLon(float gpsLon) {
-        this.gpsLon = gpsLon;
-    }
-
-
-
-    public float getGpsLat() {
-        return gpsLat;
-    }
-
-    public float getGpsLon() {
-        return gpsLon;
-    }
-
-
-}
