+++ src/main/java/eatseasyspring/eatseasyspring/service/RestaurantService.java
@@ -4,6 +4,7 @@ import eatseasyspring.eatseasyspring.repository.RestaurantRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+
 @Service
 public class RestaurantService {
     @Autowired
